Moved new coordination service types to top namespace level
