+++ src/AI4E.Coordination/CoordinationManagerOptions.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace AI4E.Coordination
+{
+    public class CoordinationManagerOptions
+    {
+        internal static TimeSpan LeaseLengthDefault { get; } = TimeSpan.FromSeconds(30);
+
+        public TimeSpan LeaseLength { get; set; } = LeaseLengthDefault;
+    }
+}
