+++ src/Common/ObjectExtension.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
+using System.Runtime.CompilerServices;
 using System.Threading.Tasks;
 using AI4E.Async;
 
@@ -44,6 +46,25 @@ namespace AI4E.Internal
         public static IEnumerable<T> Yield<T>(this T t)
         {
             yield return t;
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static T Assert<T>(this T t, Func<T, bool> assertion)
+        {
+            Debug.Assert(assertion(t));
+
+            return t;
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static T Assert<T>(this T t, Func<T, bool> precondition, Func<T, bool> assertion)
+        {
+            if (precondition(t))
+            {
+                Debug.Assert(assertion(t));
+            }
+
+            return t;
         }
     }
 }
